.github/inference-services/requirements.txt
boto3
pinecone-client
transformers
torch
faiss-cpu
fastapi
uvicorn
scipy
numpy
spacy
rdflib